Add support for specifying a remote server...
...using an incredibly nasty hack.

webRequest blocking needs a blocking response on Chromelikes, so wrap it in the Storage promise.

The downside of this is that changing the options apparently requires a full browser restart. Also, since this is a dev extension, options will probably get wiped out occasionally.

Tested and working on my desktop + SBC, though.

There's got to be a better way to do this, but as long as this works I'm not touching it.